# api/openapi.yaml
openapi: 3.1.0
info:
  title: Open Source Project Intelligence API
  version: 1.0.0
  description: Repository-owned frozen HTTP contract. Resource IDs are opaque decimal strings.
servers:
  - url: http://localhost:8100
paths:
...
  /api/v1/projects/{project_id}/contributors:
    get:
      operationId: getApiV1ProjectsProjectIdContributors
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: window
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/adoption:
    get:
      operationId: getApiV1ProjectsProjectIdAdoption
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: window
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/security:
    get:
      operationId: getApiV1ProjectsProjectIdSecurity
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: window
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/comparisons:
    post:
      operationId: postApiV1Comparisons
      tags: [contract]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '201':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/comparisons/{comparison_id}:
    get:
      operationId: getApiV1ComparisonsComparisonId
      tags: [contract]
      parameters:
        - name: comparison_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/trends:
    get:
      operationId: getApiV1ProjectsProjectIdTrends
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: kind
          in: query
          required: false
          schema:
            type: string
        - name: window
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/recommendation:
    get:
      operationId: getApiV1ProjectsProjectIdRecommendation
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: policy
          in: query
          required: false
          schema:
            type: string
        - name: window
          in: query
          required: false
          schema:
            type: string
        - name: cutoff
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/policies:
    get:
      operationId: getApiV1Policies
      tags: [contract]
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: postApiV1Policies
      tags: [contract]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '201':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/policies/{policy_id}/versions/{version}:
    get:
      operationId: getApiV1PoliciesPolicyIdVersionsVersion
      tags: [contract]
      parameters:
        - name: policy_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: version
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/policies/{policy_id}/versions:
    post:
      operationId: postApiV1PoliciesPolicyIdVersions
      tags: [contract]
      parameters:
        - name: policy_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '201':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/policies/{policy_id}/versions/{version}/activation:
    post:
      operationId: postApiV1PoliciesPolicyIdVersionsVersionActivation
      tags: [contract]
      parameters:
        - name: policy_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: version
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/radar:
    get:
      operationId: getApiV1Radar
      tags: [contract]
      parameters:
        - name: policy
          in: query
          required: false
          schema:
            type: string
        - name: window
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/radar/{project_id}/override:
    post:
      operationId: postApiV1RadarProjectIdOverride
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '201':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteApiV1RadarProjectIdOverride
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '204':
          description: Completed without a response body
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/topics:
    get:
      operationId: getApiV1ProjectsProjectIdTopics
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: window
          in: query
          required: false
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/topics/{topic_id}/corrections:
    post:
      operationId: postApiV1ProjectsProjectIdTopicsTopicIdCorrections
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: topic_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '202':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/releases:
    get:
      operationId: getApiV1ProjectsProjectIdReleases
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/releases/{release_id}:
    get:
      operationId: getApiV1ProjectsProjectIdReleasesReleaseId
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - name: release_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/crawls:
    post:
      operationId: postApiV1ProjectsProjectIdCrawls
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '202':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/knowledge/search:
    post:
      operationId: postApiV1ProjectsProjectIdKnowledgeSearch
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/projects/{project_id}/queries:
    post:
      operationId: postApiV1ProjectsProjectIdQueries
      tags: [contract]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '202':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/analysis-runs/{run_id}:
    get:
      operationId: getApiV1AnalysisRunsRunId
      tags: [contract]
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/analysis-runs/{run_id}/reruns:
    post:
      operationId: postApiV1AnalysisRunsRunIdReruns
      tags: [contract]
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '202':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/analysis-runs/{run_id}/feedback:
    post:
      operationId: postApiV1AnalysisRunsRunIdFeedback
      tags: [contract]
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '201':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/analysis-series/{series_id}/selection:
    post:
      operationId: postApiV1AnalysisSeriesSeriesIdSelection
      tags: [contract]
      parameters:
        - name: series_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '200':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/alert-rules:
    post:
      operationId: postApiV1AlertRules
      tags: [contract]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '201':
          description: Successful contract response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  schemas:
...
    Document:
      type: object
      additionalProperties: true
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string, pattern: '^[a-z][a-z0-9_]*$' }
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, request_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        code: { type: string, pattern: '^[a-z][a-z0-9_]*$' }
        instance: { type: string }
        request_id: { type: string }
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
  responses:
    Problem:
      description: A safe RFC 9457 problem response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'